Add new MAX column type; add static message types
All the metrics previously supported were counters that tick upward, but
now we want to report the size of the LLVM memory state, which goes up
and down; really, we're interested in how big it gets in each function.
Thus a MAX column takes the maximum of the metric at function entry and
rather than subtracting the final value from the starting one, and again
takes the maximum when aggregating across several function calls.